Volume: Use more appropriate types for some returned values
Disc number is changed from bool to u8, and revision is changed from
int to u16 (WADs can use all 16 bits, but discs can only use 8 bits).

+++ Source/Core/Core/CoreParameter.cpp
@@ -408,7 +408,7 @@ IniFile SCoreStartupParameter::LoadGameIni() const
 	return LoadGameIni(GetUniqueID(), m_revision);
 }
 
-IniFile SCoreStartupParameter::LoadDefaultGameIni(const std::string& id, int revision)
+IniFile SCoreStartupParameter::LoadDefaultGameIni(const std::string& id, u16 revision)
 {
 	IniFile game_ini;
 	for (const std::string& filename : GetGameIniFilenames(id, revision))
@@ -416,7 +416,7 @@ IniFile SCoreStartupParameter::LoadDefaultGameIni(const std::string& id, int rev
 	return game_ini;
 }
 
-IniFile SCoreStartupParameter::LoadLocalGameIni(const std::string& id, int revision)
+IniFile SCoreStartupParameter::LoadLocalGameIni(const std::string& id, u16 revision)
 {
 	IniFile game_ini;
 	for (const std::string& filename : GetGameIniFilenames(id, revision))
@@ -424,7 +424,7 @@ IniFile SCoreStartupParameter::LoadLocalGameIni(const std::string& id, int revis
 	return game_ini;
 }
 
-IniFile SCoreStartupParameter::LoadGameIni(const std::string& id, int revision)
+IniFile SCoreStartupParameter::LoadGameIni(const std::string& id, u16 revision)
 {
 	IniFile game_ini;
 	for (const std::string& filename : GetGameIniFilenames(id, revision))
@@ -435,7 +435,7 @@ IniFile SCoreStartupParameter::LoadGameIni(const std::string& id, int revision)
 }
 
 // Returns all possible filenames in ascending order of priority
-std::vector<std::string> SCoreStartupParameter::GetGameIniFilenames(const std::string& id, int revision)
+std::vector<std::string> SCoreStartupParameter::GetGameIniFilenames(const std::string& id, u16 revision)
 {
 	std::vector<std::string> filenames;
